Skip remeasure requests for precomposed items
When we do precomposition we add a subtree as a non placed child of SubcomposeLayout. Before those changes it was possible for this tree to be measured as part of the regular remeasure logic, as we weren't properly skipping remeasure requests from non placed items, which can't affect the placed (and visible) tree. It was a performance issue as well, as the idea of prefetch to do all the work outside of the main frame, we want to do measurement separately when we have enough free time, not during the main frame blocking the rest of the ui thread job needed to prepare the next frame.

Test: new tests in SubcomposeLayoutTest
Fixes: 371012732
(cherry picked from https://android-review.googlesource.com/q/commit:bb3785ddf9dd9576562d970d5016c07cdc35ff87)
Merged-In: Ifd9f96592b33be02338064da01a4f6b523cfe41b
Change-Id: Ifd9f96592b33be02338064da01a4f6b523cfe41b