End animations in SeekableTransitionState if already past new duration
SeekableTransitionState can get stuck with startNanoTime = 0 if an
animation is removed while it's still running. That results in the
Transition returning isRunning = true even though no animations are
playing. This guards against that by, whenever the duration changes,
checking if the new duration is shorter than the current play time, and
if so, ending the current animation.

I'm not sure if this is the best fix for this. animateOneFrameLambda is
running and it does set currentAnimation to null indicating it's
finished, but for some reason that doesn't result in startNanoTime being
reset.

Fixes: 400473060
Test: Added test with repro case from 400473060
(cherry picked from https://android-review.googlesource.com/q/commit:0c832c139dc844fcc611c7b5cdb1496ba5071a3a)
Merged-In: Ib2a3878514d490e924f8a473e2fdfc8162ad9ff5
Change-Id: Ib2a3878514d490e924f8a473e2fdfc8162ad9ff5