Make backing html input not exceed the canvas bounds and respect the dimensions of the focused rect (#2081)
(Cherry pick from
https://github.com/JetBrains/compose-multiplatform-core/pull/2070)

Fixes a second part of https://youtrack.jetbrains.com/issue/CMP-7836

The problem was that a browser tries to bring the backing html input
into a view. When the view is beyond the canvas bounds, then it would
lead to canvas being partially or completely hidden.

The solution is to not let an html input be positioned anywhere beyond
the canvas.


## Testing
Tested manually in a newly added demo-reprdocuer.

This should be tested by QA

## Release Notes
### Fixes - Web
- Fixed the positioning and the dimensions of the backing text input
(html element). The bug used to lead to unexpected scrolls on the page
due browser trying to bring the html elemtent into a view.