Change getFocusedRect() to set an empty rect when there's nothing focusable
Bug: 369256395

Relnote: "Flag ComposeUiFlags.isGetFocusedRectReturnEmptyEnabled makes it
so that when nothing is focusable in a ComposeView, it sets the rect
to an Empty value. This prevents a focus search from choosing the View
to focus on. This is especially important for when an IME has a NEXT
option where it might try to focus on the ComposeView and requestFocus()
fails."

Test: new tests, existing tests, manual testing.
(cherry picked from https://android-review.googlesource.com/q/commit:003daa8a646cdb6484ec9319114fa24932b78e8a)
Merged-In: Ibd0e260624d662b36c846bd0b2fa6e7b0f5ad08e
Change-Id: Ibd0e260624d662b36c846bd0b2fa6e7b0f5ad08e