CMP-7543 Retain AnnotatedString in ClipEntries on Desktop
Prior to this, the implementation of ClipboardUtils.desktop.kt would "flatten" AnnotatedStrings into their plaintext representation. This means that if you implemented a custom Clipboard, it would only receive the plaintext representation of the copied text in setEntry. With this change, the AnnotatedString is actually preserved when copying, so Clipboard can access the rich text and serialize it as it prefers.