Move UIKitInstrumentedTests.sh script to the repo from CI
Experiment to fix https://youtrack.jetbrains.com/issue/CMP-8136/PR-checks-runs-always-by-main-teamcity-config

If it is successful, later other PR checks will be written this way.

CI will call it this way:
```
[ -f .jb-ciscripts/PullRequests/UIKitInstrumentedTests.sh ] && .jb-ciscripts/PullRequests/UIKitInstrumentedTests.sh
```

In the release branch it doesn't exist, so it will be green. In the main branch it will run all the tests

## Release Notes
N/A