Fix caret placement near complex unicode glyphs with non-spacing marks (#2147)
Added check for a non-space mark in the getOffsetByPosition, which is
being used to determine caret position in text.
There were two problems: it was possible to manually place the caret
between parts of complex glyph, and layout calculation incorrectly
worked with non-space marks at the end of the line, placing the caret
between parts of the symbol.

Fixes: https://youtrack.jetbrains.com/issue/CMP-8054/

## Testing
This should be tested by QA

The test case is available on the related issue for iOS

Additional test case: try to click right at the `क्` character (or move
the caret by long-tap), and try to type. This symbol shouldn't be
divided by typed text

## Release Notes
### Fixes - Multiple Platforms
Fixed caret placement near glyphs if glyphs are compound symbols and
part of them are non-spacing marks