CMP-9264 Expose showLayoutBounds API on ComposePanel
This exposes a showLayoutBounds property on ComposePanel (desktop only)
that allows users to turn on or off the showLayoutBounds property on
the root layout node owner(s).

The change mostly pipes the property up to the ComposePanel, but has no
public API impact. The new property is also annotated as experimental.
The Skiko RootNodeOwner's showLayoutBounds is now also a state, to cause
a full recomposition when the value changes and see it reflected in the
UI.