CMP-9342 Fix text context menu behaviour
This fixes two issues with text context menus:

### 1. The menu should not show if it's empty
If the context menu has no items, it should not show.
This is easily achieved by only popping up the menu if
the items list is not empty.

### 2. The menu should not pop up "after the fact"
Once 1 is fixed, there is another issue that crops up:
the menu which was not popped up initially, becomes
visible as soon as there would be items in it. For
example, if no menu pops up in a `SelectionContainer`
if the user right-clicks an empty spot, but then does
a selection. Same for BTF, if the user types any text
in an initially-empty BTF after right clicking it.

This is done by explicitly closing the menu session if
it contains no items.

Note: JPopupContextMenuRepresentation works differently
and this only explicitly fixes #1 for it. Due to it
being a completely different implementation, this does
not really suffer from #2